% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitLandDF.R
\name{dims}
\alias{dims}
\title{Get Dimensions of Fitness Landscape}
\usage{
dims(x)
}
\arguments{
\item{x}{FitLandDF object}
}
\value{
integer vector analogous to `base::dim`
}
\description{
Get Dimensions of Fitness Landscape
}
\examples{
# create flat fitness landscape with dimensions 3x3x3
values <- array(0, dim = rep(3, 3))
my_landscape <- FitLandDF(values)

# print dimensions
dims(my_landscape)

}
