% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitLandDF.R
\name{extract_df}
\alias{extract_df}
\title{Extract Data Frame Representation of Fitness Landscape}
\usage{
extract_df(x)
}
\arguments{
\item{x}{FitLandDF object}
}
\value{
data frame representation of fitness landscape
}
\description{
Extract Data Frame Representation of Fitness Landscape
}
\examples{
# create fitness landscape
values <- array(1:27, dim = rep(3, 3))
my_landscape <- FitLandDF(values)

# extact data frame representation
my_df <- extract_df(my_landscape)
}
