% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitLandDF.R
\name{isFitLandDF}
\alias{isFitLandDF}
\alias{is.FitLandDF}
\alias{is_FitLandDF}
\title{Confirm Object is Valid Instance of FitLandDF}
\usage{
is.FitLandDF(x)

is_FitLandDF(x)
}
\arguments{
\item{x}{object whose class is in question}
}
\value{
`logical`; `TRUE` if `x` is an instance of FitLandDF,
  `FALSE` otherwise
}
\description{
Confirm Object is Valid Instance of FitLandDF
}
