% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitLandDF.R
\name{minmax}
\alias{minmax}
\alias{min_fit}
\alias{max_fit}
\title{Get Highest and Lowest Fitness Values from Fitness Landscape}
\usage{
min_fit(x, ...)

max_fit(x, ...)
}
\arguments{
\item{x}{FitLandDF object}
}
\value{
minimum or maximum fitness value in this landscape
}
\description{
Get Highest and Lowest Fitness Values from Fitness Landscape
}
\examples{
# create fitness landscape with min value 1 and max value 27
values <- array(1:27, dim = rep(3, 3))
my_landscape <- FitLandDF(values)

# calculate maximum fitness value
max_fit(my_landscape)

# calculate minimum fitness value
min_fit(my_landscape)

}
